move scalar_funcs and some hashing stuff into spark-expr
