+++ native/core/Cargo.toml
@@ -58,7 +58,7 @@ brotli = "3.3"
 flate2 = "1.0"
 lz4 = "1.24"
 zstd = "0.11"
-rand = "0.8"
+rand = { workspace = true}
 num = { workspace = true }
 bytes = "1.5.0"
 tempfile = "3.8.0"
@@ -70,7 +70,6 @@ datafusion = {  workspace = true }
 datafusion-expr = {  workspace = true }
 datafusion-physical-expr-common = {  workspace = true }
 datafusion-physical-expr = {  workspace = true }
-unicode-segmentation = "^1.10.1"
 once_cell = "1.18.0"
 regex = { workspace = true }
 crc32fast = "1.3.2"
@@ -85,7 +84,6 @@ jni = { version = "0.21", features = ["invocation"] }
 lazy_static = "1.4"
 assertables = "7"
 hex = "0.4.3"
-twox-hash = "1.6.3"
 
 [features]
 default = []
