+++ native/core/src/execution/datafusion/mod.rs
@@ -21,5 +21,4 @@ pub mod expressions;
 mod operators;
 pub mod planner;
 pub mod shuffle_writer;
-pub mod spark_hash;
 mod util;
