+++ native/spark-expr/Cargo.toml
@@ -40,6 +40,12 @@ chrono-tz = { workspace = true }
 num = { workspace = true }
 regex = { workspace = true }
 thiserror = { workspace = true }
+unicode-segmentation = "1.11.0"
+
+[dev-dependencies]
+arrow-data = {workspace = true}
+rand = { workspace = true}
+twox-hash = "1.6.3"
 
 [lib]
 name = "datafusion_comet_spark_expr"
