+++ native/spark-expr/src/scalar_funcs.rs
