+++ native/spark-expr/src/scalar_funcs/chr.rs
