+++ native/spark-expr/src/scalar_funcs/hex.rs
